Minor changes to surpress langchain-community warnings and deprication warnings
