+++ utils/document_handler.py
@@ -1,7 +1,7 @@
 import os
 import pathlib
 
-from langchain.document_loaders import DirectoryLoader
+from langchain_community.document_loaders import DirectoryLoader
 from langchain.text_splitter import RecursiveCharacterTextSplitter
